requirements.txt
atlassian-python-api==4.0.4
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
Deprecated==1.2.18
idna==3.10
jmespath==1.0.1
markdown2==2.5.3
oauthlib==3.2.2
python-dotenv==1.1.0
requests==2.32.4
requests-oauthlib==2.0.0
soupsieve==2.7
typing_extensions==4.14.0
urllib3==2.4.0
wrapt==1.17.2
